:bug: Fix: payment price 관련 타입 로직 수정

+++ src/main/java/com/ohsooo/platform/ohsooshoppingmall/domain/payment/dto/response/RefundDetailResponseDto.java
@@ -1,6 +1,7 @@
 package com.ohsooo.platform.ohsooshoppingmall.domain.payment.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -23,7 +24,7 @@ public class RefundDetailResponseDto {
   private String status;
 
   @Schema(description = "환불 금액", example = "5000")
-  private Integer amount;
+  private BigDecimal amount;
 
   @Schema(description = "환불 사유(선택)", example = "단순 변심")
   private String reason;
@@ -44,7 +45,7 @@ public class RefundDetailResponseDto {
   public static class RefundItem {
     private Long refundItemId;
     private Long orderItemId;
-    private Integer amount;
+    private BigDecimal amount;
     private Integer quantity;
   }
 }
